moved assets from src to public for better accessing
